+++ constants/widget-editor.js
@@ -1,0 +1,6 @@
+export const WIDGET_EDITOR_DEFAULT_DISABLED_FEATURES = [
+  'typography',
+  'end-user-filters',
+];
+
+export default { WIDGET_EDITOR_DEFAULT_DISABLED_FEATURES };